+++ src/Explosion.h
@@ -17,7 +17,7 @@ class Explosion : public Movable
 public:
 
     // The dynamite uses this 
-    Explosion(const Posf& posf, const Posb& posb, int radius, Color color, 
+    Explosion(const Posf& posf, const Posb& posb, size_t radius, Color color, 
         ExplosionsBoardProxy& explosionProxy);
 
 
@@ -51,7 +51,7 @@ private:
 
     ExplosionsBoardProxy& _explosionProxy;
 
-    int _radius;
+    size_t _radius;
 
     bool _propagate = false;
     bool _center = false;
@@ -67,7 +67,7 @@ private:
     // only an explosion can create another explosion. 
     // The first explosion will create 4 different explosions
     // the rest of them 
-    Explosion(const Posf& posf, const Posb& posb, int radius, Color color, 
+    Explosion(const Posf& posf, const Posb& posb, size_t radius, Color color, 
         Movable::Direction dir, ExplosionsBoardProxy& explosionProxy);
 
     //Posb getNextTile() const;
